requirements.txt
gradio==3.39.0
insightface==0.7.3
onnxruntime==1.23.1
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.3.0
websockets<12.0
